+++ src/main/java/com/atlauncher/Data.java
@@ -22,7 +22,6 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
-import com.atlauncher.data.AbstractAccount;
 import com.atlauncher.data.LWJGLVersions;
 import com.atlauncher.data.News;
 import com.atlauncher.data.Pack;
@@ -31,8 +30,6 @@ import com.atlauncher.data.minecraft.JavaRuntimes;
 import com.atlauncher.data.minecraft.VersionManifestVersion;
 
 public final class Data {
-    public static final List<AbstractAccount> ACCOUNTS = new LinkedList<>();
-    public static AbstractAccount SELECTED_ACCOUNT = null; // Account using the Launcher
 
     public static Map<String, Object> CONFIG = new HashMap<>();
     public static Map<String, Object> CONFIG_OVERRIDES = new HashMap<>();
@@ -40,8 +37,6 @@ public final class Data {
     public static final List<News> NEWS = new LinkedList<>();
 
     public static final List<Pack> PACKS = new LinkedList<>();
-
-    public static final List<Server> SERVERS = new LinkedList<>();
 
     public static final Map<String, VersionManifestVersion> MINECRAFT = new HashMap<>();
     public static LWJGLVersions LWJGL_VERSIONS = null;