+++ src/main/java/com/atlauncher/gui/card/packbrowser/UnifiedPackCard.java
@@ -101,7 +101,7 @@ public class UnifiedPackCard extends JPanel implements RelocalizationListener {
                         .setType(DialogManager.ERROR).show();
 
                 if (AccountManager.getAccounts().size() == 0) {
-                    App.launcherFrame.tabbedPane.setSelectedIndex(UIConstants.LAUNCHER_ACCOUNTS_TAB);
+                    App.navigate(UIConstants.LAUNCHER_ACCOUNTS_TAB);
                 }
             } else {
                 Analytics.trackEvent(AnalyticsEvent.forPackInstall(result));
@@ -130,7 +130,7 @@ public class UnifiedPackCard extends JPanel implements RelocalizationListener {
                         .setType(DialogManager.ERROR).show();
 
                 if (AccountManager.getAccounts().size() == 0) {
-                    App.launcherFrame.tabbedPane.setSelectedIndex(UIConstants.LAUNCHER_ACCOUNTS_TAB);
+                    App.navigate(UIConstants.LAUNCHER_ACCOUNTS_TAB);
                 }
             } else {
                 Analytics.trackEvent(AnalyticsEvent.forPackInstall(result, true));