+++ src/main/java/com/atlauncher/gui/dialogs/InstanceInstallerDialog.java
@@ -457,7 +457,19 @@ public class InstanceInstallerDialog extends JDialog {
                 List<CurseForgeFile> serverFiles = CurseForgeApi.getFiles(serverFileIds);
 
                 dialog.setReturnValue(
-                        serverFiles.stream().filter(f -> f.isAvailable && f.isServerPack && f.getGameVersion() != null)
+                        serverFiles.stream().map(f -> {
+                            if (f.getGameVersion() == null) {
+                                Optional<CurseForgeFile> matchingFile = files.stream()
+                                        .filter(sf -> sf.serverPackFileId != null)
+                                        .filter(sf -> sf.serverPackFileId == f.id).findFirst();
+
+                                if (matchingFile.isPresent()) {
+                                    f.gameVersions = matchingFile.get().gameVersions;
+                                }
+                            }
+
+                            return f;
+                        }).filter(f -> f.isAvailable && f.isServerPack && f.getGameVersion() != null)
                                 .collect(Collectors.toList()));
             } else {
                 dialog.setReturnValue(files);
@@ -662,24 +674,26 @@ public class InstanceInstallerDialog extends JDialog {
     }
 
     private void handleTechnicInstall(Object object) {
+        String slug;
+
         if (object instanceof TechnicModpack) {
-            technicModpack = (TechnicModpack) object;
+            slug = ((TechnicModpack) object).name;
         } else {
-            String slug = ((TechnicModpackSlim) object).slug;
-
-            final ProgressDialog<TechnicModpack> technicModpackDialog = new ProgressDialog<>(
-                    GetText.tr("Getting Modpack Details"), 0, GetText.tr("Getting Modpack Details"),
-                    "Aborting Getting Modpack Details");
-
-            technicModpackDialog.addThread(new Thread(() -> {
-                technicModpackDialog.setReturnValue(TechnicApi.getModpackBySlug(slug));
-
-                technicModpackDialog.close();
-            }));
-
-            technicModpackDialog.start();
-            technicModpack = technicModpackDialog.getReturnValue();
+            slug = ((TechnicModpackSlim) object).slug;
         }
+
+        final ProgressDialog<TechnicModpack> technicModpackDialog = new ProgressDialog<>(
+                GetText.tr("Getting Modpack Details"), 0, GetText.tr("Getting Modpack Details"),
+                "Aborting Getting Modpack Details");
+
+        technicModpackDialog.addThread(new Thread(() -> {
+            technicModpackDialog.setReturnValue(TechnicApi.getModpackBySlug(slug));
+
+            technicModpackDialog.close();
+        }));
+
+        technicModpackDialog.start();
+        technicModpack = technicModpackDialog.getReturnValue();
 
         pack = new Pack();
         pack.externalId = technicModpack.id;