+++ src/test/java/ui/AbstractUiTest.java
@@ -81,7 +81,7 @@ public class AbstractUiTest extends AssertJSwingTestCaseTemplate {
         // start the application
         ApplicationLauncher.application(App.class)
                 .withArgs("--skip-setup-dialog", "--disable-analytics", "--disable-error-reporting",
-                        "--skip-tray-integration", "--no-launcher-update", "--disable-http2", "--proxy-type=SOCKS",
+                        "--skip-tray-integration", "--no-launcher-update", "--proxy-type=SOCKS",
                         "--proxy-host=127.0.0.1", "--proxy-port=" + mockServer.getPort(),
                         "--working-dir=" + workingDir.toString())
                 .start();